Add filters and rename request
