+++ controllers/index.js
@@ -74,7 +74,7 @@ router.use('/decoders', require('./decoders'));
 router.use('/cache', require('./cache'));
 router.use('/cluster', require('./cluster'));
 router.use('/syscollector', require('./syscollector'));
-router.use('/unique', require('./unique'));
+router.use('/report', require('./report'));
 
 // Index
 router.get('/',function(req, res) {
